Install bash completion files in writable directory
Snapd installs bash completion files from snaps in
`/usr/share/bash-completion/completions` which in some distributions
is a read-only filesystem.

Instead of installing them in `/usr` we can install them within
`/var/lib/snapd` which should always be writable.

To enable that path in bash completions we only need to add the path
to `XDG_DATA_DIRS`.

An alternative could be to install the file in
`/etc/bash_completion.d`.